Add a button to manually send the Preflight Check message to the rocket. Verify communication to the rocket and tower (correct addresses and ports).
git-svn-id: http://svn.psas.pdx.edu/svn/psas/trunk/java@1043 6115d4fc-9e01-0410-8836-dad5f2c52705

+++ src/cansocket/CanMessage.java
@@ -45,9 +45,9 @@ public class CanMessage extends NetMessage
     /* Construct a can message from given id, timestamp, body.
      * This is the packed 16-bit id containing (id11,rtr,len)
      */
-    public CanMessage(short id16, int timestamp, byte body[])
+    public CanMessage(int id16, int timestamp, byte body[])
     {
-	this.id = id16;
+	this.id = (short)id16;
 	this.timestamp = timestamp;
 	this.body = body;
     }
